+++ python_jamf/server.py
@@ -4,6 +4,7 @@ from jamf_auth import (
     JamfAuthException,
 )
 from jps_api_wrapper.classic import Classic
+from jps_api_wrapper.pro import Pro
 
 from . import config, exceptions, records
 
@@ -43,8 +44,10 @@ class Server:
                 "No jamf hostname or credentials could be found."
             )
         self.classic = None
+        self.pro = None
         self.debug = debug
-        self.records = records
+        self._context_id = f"server-{id(self)}"
+        self._records_cache = {}
         try:
             self.classic = Classic(
                 self.config.hostname,
@@ -52,14 +55,36 @@ class Server:
                 self.config.password,
                 client=self.config.client,
             )
+            self.pro = Pro(
+                self.config.hostname,
+                self.config.username,
+                self.config.password,
+                client=self.config.client,
+            )
         except AuthResponseConnectionError:
             print("Could not connect to " + self.config.hostname)
-            exit()
+            raise
         except JamfAuthException:
             print("Incorrect username or password for " + self.config.hostname)
-            exit()
+            raise
         except AuthResponseWasNotValid:
             print("Some error for " + self.config.hostname)
-            exit()
-        self.records.set_classic(self.classic)
-        self.records.set_debug(debug)
+            raise
+
+    def record_class(self, name, case_sensitive=False):
+        return records.class_name(name, case_sensitive=case_sensitive)
+
+    def records(self, record_cls):
+        if isinstance(record_cls, str):
+            record_cls = self.record_class(record_cls, case_sensitive=False)
+        if record_cls not in self._records_cache:
+            self._records_cache[record_cls] = record_cls(
+                classic=self.classic,
+                #pro=self.pro,
+                debug=self.debug,
+                context_id=self._context_id,
+            )
+        return self._records_cache[record_cls]
+
+    def records_by_name(self, name, case_sensitive=False):
+        return self.records(self.record_class(name, case_sensitive=case_sensitive))